Fix our IsHealthy expectations during candidate promotion.
Both states REPORT_LSN and PREPARE_PROMOTION are expected to happen without
Postgres running, so we should accept to find a node that is being promoted
but is not marked reachable/healthy on the monitor yet.

We want to be able to select a new candidate if the previously selected
candidate was not available anymore after selection, but we also want to
avoid having two candidates at the same time.

In this fix we also add some more messages on the monitor to make it obvious
what's happening with finding a failover-in-progress:

    17:32:31 87 INFO  New state for node 1 "node_1" (172.27.1.3:5432): join_secondary ➜ join_secondary
    17:32:31 87 INFO  Active node 1 "node_1" (172.27.1.3:5432) found failover candidate node 2 "node_2" (172.27.1.4:5432) being promoted, it reported state "prepare_promotion"
    17:32:31 87 INFO  Active node 3 "node_3" (172.27.1.5:5432) found failover candidate node 2 "node_2" (172.27.1.4:5432) being promoted, it reported state "prepare_promotion"
    17:32:31 87 INFO  Active node 2 "node_2" (172.27.1.4:5432) found failover candidate node 2 "node_2" (172.27.1.4:5432) being promoted, it reported state "prepare_promotion"
    17:32:31 87 INFO  Active node 1 "node_1" (172.27.1.3:5432) found failover candidate node 2 "node_2" (172.27.1.4:5432) being promoted, it reported state "prepare_promotion"
    17:32:31 87 INFO  node 2 "node_2" (172.27.1.4:5432) reported new state "stop_replication"