Add support for running a pg_autoctl controlled pgbouncer instance
Implements pg_autoctl [enable|disable] pgbouncer [options]
Implements pg_autoctl create postgres [--enable-pgbouncer <file>]

Intention:
	In a pg_autoctl managed environment, have a pgbouncer instance pointing to
the current primary. The instance can run in any postgres node but not in a
monitor node. The instance can start either when the postgres node does or at
any other moment. The instance can stop at any moment without affecting the
running node.

Architecture:
	Process tree:
		In essence we want to run pgbouncer as a subprocess of pg_autoctl. It will
	be pg_autoctl's responsibility to manage the pgbouncer subprocess. In that
	effect, pg_autoctl launches a 'pgbouncer manager' process and a pgbouncer
	process under it. When running, the process tree may look like this:

	    pg_autoctl create postgres --pgdata ./data ---
	     \_ pg_autoctl: start/stop postgres
	     |   \_ /postgres -D ./data -h *
	     |       \_ postgres: logger
	     |       \_ postgres: startup recovering 000000030000000000000003
	     |       \_ postgres: checkpointer
	     |       \_ postgres: background writer
	     |       \_ postgres: stats collector
	     |       \_ postgres: walreceiver streaming 0/305FB48
	     \_ pg_autoctl: node active
	     \_ pg_autoctl: pgbouncer manager
	         \_ pgbouncer config.ini

	The top-level pg_autoctl process needs not to know about the pgbouncer
	process. It only needs to know about the 'pgbouncer manager' process. This
	is equivelant with relation between top level pg_autoctl,  'start/stop
	postgres' and postgres itself.

	Pgbouncer manager:
		This service is responsible for starting,stoping,pausing and resuming
	pgbouncer. Also, it listens to notifications from the monitor node regarding
	the state of the nodes and upon a state change, it pauses pgbouncer, updates
	its running configuration (more about configuration in the section
	'Configuration') and finally resumes pgbouncer.

	The service will exit either when asked to from the top-level pg_autoctl
	process, or if pgbouncer fails. It is the top level's process responsibility
	to restart the pgbouncer manager.

	Keeper:
		Intentionally, keeper does not keep any state regarding both pgbouncer
	manager and pgbouncer itself. The rationale for this is that pgbouncer
	should be considered a 'best effort' service. One should explicitly point
	that the service is required on every startup. If pgbouncer fails, one
	should not lose failover requirements of the database. Of course, one can
	enable/disable the pgbouncer service when one wishes.

Configuration:
	Pgbouncer is heavily configuration based. It is that no command line
arguments will be used and everything will be explicitly written in the
configuration file that is finally passed as an argument to pgbouncer.

	When the user of pg_autoctl requires a pgbouncer service, then she has to
point to a pgbouncer configuration file. At that point, pg_autoctl takes control
of the configuration. It will only read the 'user' and 'pgbouncer' sections and
ignore the 'database' section. Also, it will not read all the keys from those
two sections, because the pgbouncer set up has to match the set up of the nodes.

	Once the keys are read, then they are written to a special template file
handled by pg_autoctl entirely. Any keys that are known to point to files, will
be parsed and the contents of said files will be copied over to our template
directory. Whenever a pgbouncer process needs to read the configuration, the
template file will be copied over to a special run time configuration file. Then
any run time keys will be added, eg. logfile location, pid file location, etc,
along with the database section.

	The database section contains a single entry only, pointing to the
currently running primary node.

	All configuration file management is implemented following the XDG Base
Directory specification. See the structure ConfigFilePaths and
pgbouncer_config{.c,.h} for details.

	Intentionally no template files are cleaned up at process exit. The same
applies for the runtime files although the last is more for debugging purposes
at the moment and should be addressed in the future.

	Implementatio detail:
		Given the extent of the configuration options of pgbouncer, the
structure that is holding all that info is best to not be kept in the stack. We
are not responsible for validation of the options, yet it is easier to map the
type of the options to either strings or integers. Since the ini api used in
pg_auto_failover codebase initializes integer values to -1, a section has been
added to skip integer keys with that value when writing ini files. In retrospect
having -1 as the default value seems not a great choice as it is not uncommon
for -1 to mean "explicitly disable something" as opposed to "use default value".
It might be better to use INT_MIN for the default initializer of integers and
skip that value if needed.

NOTE: At this point, not all keys are handled (e.g. ssl, auth_type etc) in order
to get the full picture first straight before adding those trivial issues.

Supervisor:
	It was desired to use the already extensive supervisor infrastructure of
pg_autoctl to run the pgbouncer manager process under. It was also desired to be
able to start or stop the pbouncer manager process without affecting any other
running core services, such as the database. Also, it was desired for the keeper
service to not track any state regarding pgbouncer, because as explained above,
a failing pgbouncer should not lead to any node stop.

	For this, a new concept of 'enabled' and 'enableable' service is added. The
supervisor can include services that can be enabled during startup or at any
other point after the startup. Struct Service is expanded with two new members,
'enabled' to denote the current running state of service which defaults to yes,
and 'canDisable' to denote if we can ever disable that service which defaults to
'no'. See supervisor.h for details. When Service is declared with the
'canDisable' flag to 'true' it is possible to enable it or disable it.

	Two new signals have been added to handle 'enable' and 'disable' cases,
'USR1' and 'USR2' respectively. Upon receipt of an 'enable' signal, the
supervisor goes round its services ring and looks for services that can be
enabled, if found, it's start_service function is called and on successfull
completion the service is marked as enabled. Upon receipt of a 'disable' signal,
the supervisor goes round its services ring and looks for enabled services that
can be disabled, if found then the supervisor sends the stop service signal to
it and marks it as 'disabled'.

	A service that is marked as 'canDisable', can also have any policy defined.
If an enabled service fails the supervisor fails to restart it in case it is
permanent, the supervisor should not stop all the other services running.

	For pgbouncer manager, the policy is set as PERMANENT as it is desired for
the service to keep running as long as possible.

Enable/Disable
	It follows that pg_autoctl [enable|disable] pgbouncer [options] can be
implemented simply by writing the configuration template and signaling the
running node to enable or disable its services. If the command was 'enable', it
would be preferable and trivial to issue a 'select 1' command to pgbouncer to
ensure that it came up. This is left for a subsequent commit.

Signaling:
	At this point, the communication between pgbouncer manager and pgbouncer is
happening with signals. It will be preferable for this communication to happen
using psql commands. It is rather trivial but it adds quite a bit to the already
large diff footprint. For this, it is left for a subsequent commit

	If a more fine-grained control over which services are enabled (since now is
start everything/stop everything implementation) is desired, then signalling is
not the way to go. This is left for a subsequent commit.