pg_autoctl do azure create
Review commands so that the main entry point for azure QA setup creates all
the VMs, provision them with our selected software, and then create a tmux
session with interactive control over the setup. The tmux session is primed
with commands that deploy pg_autoctl on each node, including integration in
the systemd setup.