Simplify FSM by moving transition to primary to a common place.
The code is lending itself to having a single place to decide if the primary
should transition to WAIT_PRIMARY or JOIN_PRIMARY or PRIMARY. The accounting
of the candidatePriority == 0 nodes is done there in a loop and should not
be replicated elsewhere.

Where we are moving a node to SECONDARY, we can wait until the primary shows
up again to realise it needs to move to PRIMARY, we don't need to already
assign the goal State.