Fix previous idea: take wait_primary in account.
In ProceedGroupStateForPrimaryNode in our loop over the secondary states to
count failover candidates and what not, we added the management of the case
when the primary node is wait_primary.

We certainly still don't want to promote from wait_primary to primary when
losing (yet another?) standby candidate node, though. So put that part of
the code back like before.