Fix pg_autoctl enable|disable maintenance waiting.
The command used to wait until one node in the group has reached primary,
though when having more than one standby node we can go to maintenance and
back without impacting the state of the primary node.

We should wait for the node that is going to maintenance to reach its own
target state, and that's what this commit does.