Refrain from editing synchronous_standby_names when in maintenance.
After all we use ANY (node1, node2) so when node2 is in maintenance there is
no harm in still having it listed in synchronous_standby_names. The only bad
thing that could happen is if the node in maintenance would ACK a
transaction and then lose it locally, I suppose, but I am not sure how to
make that happen.