Improve IsHealthy() checks.
To avoid some strange race conditions, don't wait until the next background
check changed the NODE_HEALTH_BAD to NODE_HEALTH_GOOD when the node-active
protocol is reporting that Postgres is running on the client side.

Trust the client-side node-active protocol when its most recent report is
within the client-side frequency, meaning it reported in the last second,
and the health check didn't happen more recently than that.