Attempt at fixing test_enable_ssl flappy test.
When we stop node1 (primary) the monitor might notice right away if the
background check has been scheduled at this time. To avoid this, update
node1 SSL setup without causing a restart of the node.

This time we still have to sleep for 6s and hope that the change has made it
the background process in that timing. With some luck we will have a less
flappy test that way, though.