Introduce dynamic services and a tighter integration of pgbouncer to keeper
This commit is removing the concept of 'enabled' services and introduces dynamic
services in its place.

Prior Services could be disabled or enabled based on their properties. That made
some services different than others but left most of the rest of the code
intact, with the exception of the supervisor. The service array was statically
defined and attached to the supervisor. Then the supervisor would read the
properties of each service and decide how to act in different scenarios.

Now all services look the same. The concept of being able to turn on or off
services is still maintained. However now, the responsability is split between
the caller of the supervisor and the supervisor itself. There exists a static
array of services under supervision which is declared when isntructing the
supervisor to start. If any of these services fails, the whole node fails.

Enter dynamic services. These are services that can be added and removed at any
point. If any of these services fails, then only the failed services are
affected, leaving the rest of the node intact.

These services are added in their own special array, in an anonymous struct
within the supervisor. Both arrays, comprise the set of services under
supervision.

Services can be both added and removed from that array. The responsibility to
add services does not belong to the supervisor. For this the supervisor exposes
a handler function which the caller can set during the start up. It also exposes
two functions supervisor_add_dynamic, supervisor_remove_dynamic which allow to
the caller to add or remove dynamic services.

If a dynamic service fails and can no longer be restared, then the supervisor
simply removes it from the array of dynamic services and does not inform the
caller of the supervisor for the fact. Mostly because it seems that the job of
the supervisor is to supervise.

Keeper configuration changes.

Pgbouncer runs in a keeper node, so it makes some sense for the keeper to keep
that information in its configuration. A new section has been added, 'services'.
A new function to reread that section has also been added in keeper,
keeper_reread_services. This is mostly because the rest of the read/reload
exposed functions are very tightly coupled with what can be possibly happening
with parts of the configuration and are rightfully opinionated about it.
Instead, rereading of the services is intented to be more lightweight.

For the same reason, the action to set those values from the clients, is not
following the get/set client as it explicitly calls for a reload to happen after
a value is written. We explicitly want to avoid that reload. Sure an extra
argument to skip that part could have been added, but keeping the concept
separate seemed cleaner.

Now pgbouncer is declared as a dynamic service, using the newly introduced
infrastructure. However, since it is more dependant to the keeper configation
its start up function had to be modified slightly. It now requires that the
pg_setup is ready before it starts, otherwise it will fail. During startup, it
spins around pg_setup_is_ready for a bit and if it is not, then it fails.