Review consequences of pg_autoctl set formation number-sync-standbys 0. (#535)
* Review consequences of pg_autoctl set formation number-sync-standbys 0.

In particular setting the number-sync-standbys to zero means that when we
don't have a single healthy secondary that is also a candidate for
promotion, we are allowed to reach WAIT_PRIMARY.

As a consequence, we also want to allow changing the number-sync-standbys
setting when in WAIT_PRIMARY so as to allow blocking writes when it's all
going south, in order to limit potential data loss.

* Fix tests expectations about synchronous_standby_names in wait_primary.

We can actually be in WAIT_PRIMARY for having lost all our candidate standby
nodes, and still maintain a non-empty synchronous_standby_names if we have
number_sync_standbys > 0.

* Help make obvious which node is "it" in the notification message.

* Fix race condition detection.

Two states are candidate to the race condition problem in detecting when
we're losing our standby nodes: REPORT_LSN and JOIN_SECONDARY, where in both
those states the secondary Postgres instance is not restarted yet.

* Clarify notification message.

* Simplify FSM by moving transition to primary to a common place.

The code is lending itself to having a single place to decide if the primary
should transition to WAIT_PRIMARY or JOIN_PRIMARY or PRIMARY. The accounting
of the candidatePriority == 0 nodes is done there in a loop and should not
be replicated elsewhere.

Where we are moving a node to SECONDARY, we can wait until the primary shows
up again to realise it needs to move to PRIMARY, we don't need to already
assign the goal State.

* Fix previous idea: take wait_primary in account.

In ProceedGroupStateForPrimaryNode in our loop over the secondary states to
count failover candidates and what not, we added the management of the case
when the primary node is wait_primary.

We certainly still don't want to promote from wait_primary to primary when
losing (yet another?) standby candidate node, though. So put that part of
the code back like before.

* Fix pg_autoctl enable|disable maintenance waiting.

The command used to wait until one node in the group has reached primary,
though when having more than one standby node we can go to maintenance and
back without impacting the state of the primary node.

We should wait for the node that is going to maintenance to reach its own
target state, and that's what this commit does.

* Refrain from editing synchronous_standby_names when in maintenance.

After all we use ANY (node1, node2) so when node2 is in maintenance there is
no harm in still having it listed in synchronous_standby_names. The only bad
thing that could happen is if the node in maintenance would ACK a
transaction and then lose it locally, I suppose, but I am not sure how to
make that happen.

* Leave most of the work to ProceedGroupStateForPrimaryNode.

* Fix some more tests expectations for synchronous_standby_names.

In WAIT_PRIMARY we have ssn = '', always, that's the point of assigning the
state to WAIT_PRIMARY: async replication is forced, to cope with the
situation.

* Improve notification messages using present tense.

* Remove a TODO item about removing a notification: we should keep it.

* Improve IsHealthy() checks.

To avoid some strange race conditions, don't wait until the next background
check changed the NODE_HEALTH_BAD to NODE_HEALTH_GOOD when the node-active
protocol is reporting that Postgres is running on the client side.

Trust the client-side node-active protocol when its most recent report is
within the client-side frequency, meaning it reported in the last second,
and the health check didn't happen more recently than that.

* Add unit tests for the two issues we fix in this PR.

* Attempt at fixing test_enable_ssl flappy test.

When we stop node1 (primary) the monitor might notice right away if the
background check has been scheduled at this time. To avoid this, update
node1 SSL setup without causing a restart of the node.

This time we still have to sleep for 6s and hope that the change has made it
the background process in that timing. With some luck we will have a less
flappy test that way, though.

* reindent Python code for tests.

* Per review.

Simplify things down some more, review comments, and make it so that after
APPLY_SETTINGS we pick the right choice of WAIT_PRIMARY or PRIMARY. The
waitStandbyNodeCount should not have an impact there.

* Forgot to edit a constant to use new variable...

* Review APPLY_SETTINGS to PRIMARY transition.

Now that we enable this transition in case when the formation number sync
standbys is edited while the primary node is in WAIT_PRIMARY, we must apply
the same transition function as when going from WAIT_PRIMARY to PRIMARY.

* Review indentation.

* Per review, add a test that blocks writes on a wait_standby.

By doing pg_autoctl set number-sync-standbys 1 on a wait_primary node that
has lost its standby nodes while it was with number-sync-standbys 0, we can
re-install sync rep in a write-blocking way during a production outage,
probably to bound the potential data loss in case the primary would be lost
now.