make NODES=3 NODES_ASYNC=2 NODES_PRIOS=50,0 NODES_SYNC_SB=0 cluster
This PR allows to setup an interactive testing/QA cluster with tmux with
full specifications of nodes replication settings. Before that the code was
conflating the replicationQuorum and the candidatePriority of nodes, and
that was limiting our testing abilities.

When unspecified, nodes are all created with --candidate-priority 0. Then it
is possible to specify one or more candidate priorities, which are used in
the order given. When more nodes are created than priorities are given, we
use the last specified priority again and again.

For instance --node-priorities 50,0 --nodes 3 results in using 50,0,0 as the
respective priorities for the nodes thus created.