We need either postgresql-server-dev-11 or libpq-dev.
Trouble is, when using libpq-dev, we then have /usr/bin/pg_config that
matches the installed Postgres version (in our controled case, might not be
true in general), though we don't have /usr/lib/postgresql/bin/11/pg_config
where we would expect it from /usr/lib/postgresql/bin/11/pg_ctl.

So we introduce some code that goes from the path to pg_ctl and figures out
where the pg_config might well be. When postgresql-server-dev-11 is
installed, this is not needed.

Most production environment will only install libpq-dev though, and that's
what we do also in our Dockerfile, so that we skip the whole build chain.