Feature/docker run make cluster (#556)
* Review and simplify tmux environment management.

Use tmux set-environment in the top-level tmux script, and then eval $(tmux
show-environment -s) in each window that we create with split-window.

* Review Dockerfile.

Update to current debian testing package list, which now uses Postgres 13
instead of the previous Postgres 12.

Make it possible to run the following scenario:

  $ docker build -t pg_auto_failover .
  $ docker run --name pg_auto_failover -ti --rm pg_auto_failover pg_autoctl do tmux session

* Switch our Dockerfile base OS back to debian stable.

In debian testing the Python pyroute2 librairy has changed in ways that are
not compatible with our test suite module for network abstractions:

  https://docs.pyroute2.org/ipdb_toc.html

  Warning The IPDB module is deprecated and obsoleted by NDB. Please
  consider using NDB instead.

At the moment, though, documentation is sparse and I failed to find any
guide to help with the code migration from IPDB to NDB.

* Split the Dockerfile in two containers: build-test, and "run".

The run container is expected to be public and shared with developpers who
want to try pg_auto_failover without having to install anything. We copy
files from the build-test environment where we build the full image we need
to run our unit test suite.

* Reduce/optimize Docker container size for "run".

* We need either postgresql-server-dev-11 or libpq-dev.

Trouble is, when using libpq-dev, we then have /usr/bin/pg_config that
matches the installed Postgres version (in our controled case, might not be
true in general), though we don't have /usr/lib/postgresql/bin/11/pg_config
where we would expect it from /usr/lib/postgresql/bin/11/pg_ctl.

So we introduce some code that goes from the path to pg_ctl and figures out
where the pg_config might well be. When postgresql-server-dev-11 is
installed, this is not needed.

Most production environment will only install libpq-dev though, and that's
what we do also in our Dockerfile, so that we skip the whole build chain.

* Use --privileged instead of individual Docker capabilities.

The newer Docker version seems to require that setting to work with our
requirements.

See #491.

* Per review.