Refrain from preventing changes that are not really changes.
When a user runs the same command twice in a row, such as with the
following, make it so that the second time just works. It didn't because we
failed to observe that setting the candidate priority to zero could be a
non-changer if that's already the current value of the setting.

  pg_autoctl set node candidate-priority --name node3 0

We could also skip the operation entirely, but this has two drawbacks:

  - the client implementation would then need to decide if it has to wait
    for the apply-settings steps to happen or immediately return,

  - more importantly, if there's a problem with synchronous_standby_names it
    might be good that setting a value to its current setting, thus applying
	no changes, forces a cache invalidation of the primary's setup.