Implement pg_autoctl create postgres --pg-hba-lan option. (#561)
* Implement pg_autoctl create postgres --pg-hba-lan option.

With this option pg_autoctl edits the pg_hba.conf for Postgres to grant
connection privileges on the detected LAN for the --dbname database and for
the --username user.

The LAN detection is done the same way as with the monitor.

* Implement a retry loop when trying to figure local IP address.

We connect to the monitor to figure out the local hostname and IP address,
and the monitor might not be running yet. In that case, we might want to
persist with some retries before failing back to the user.

* Install HBA on the monitor early.

This avoids some retry attemps from the other nodes at first startup.

* When we WARN about hostname, use IP address in HBA.

Rather than adding an hostname that we know faulty in the HBA file, we add
one of the IP addresses of the hostname instead. We might want to revisit
this (add all IP addresses maybe? or find the one we want to add by
connecting, like we do for the monitor?), but it allows the docker-compose
demo to just work with a minimum of trouble.

* Add a docker-compose.yml setup.

This allows to run a demo where the monitor and the Postgres nodes are each
running in their own container and connecting through the docker provided
TCP/IP network.