Minor refactor as per PR discussions
Introduce a new type for an array of services named ServiceArray and use it
throughout.

Include an example of actual usage in keeper_reread_services() which is not
included in the compiled binary though.

Rename supervisor_dynamic_service_[enable|disable] to
supervisor_[enable|disable]_dynamic_service

Include the pid of an enabled dynamic service in the log message.

Rename MAXDYNSERVICES to MAX_SERVICES since it now applies to all kinds of
services.

Be explicit when branching over a kill() call.