Remove responsibility of dynamic services from the caller of the supervisor
Instead a newly introcuded file, services.cfg, is holder of that information.
The supervisor, if permitted by the caller to handle dynamic services, it will
read that file and enable or disable services accordingly. Services are still
discoverable by name, which must be unique.

Service gets extened with a new entry, role, which is used as unique indentifier
of the service definition. A role may exist with many different names under
supervistion.

The services configuratioon file, is handled by the clients. Currently there are
no clients that need to do that, however the code is included here for
completion.