Plug minimal memleak in keeper's configuration (#583)
Admittedly this leak could have been plugged with a couple of well placed calls
to the now obsolete keeper_config_destroy() function. However, given the
lifecycle and globality of the struct the variable is found, ownership is
blurred in some cases.

Instead the leaking variable is now living in the stack along with the rest. A
new variable for its length is added instead of MAXCONNINFO which seems a bit
too excessive. Alignment concerns aside, some relief from the stack stress might
be desirable.

Finally, the necessity for keeper_config_destroy is now removed. Thus the
function itself is removed for the benefit of the reader of the codebase.

Co-authored-by: Georgios Kokolatos <gkokolatos@pm.com>