Implement a demo application showing client-side HA with reconnections (#568)
* Install HBA on the monitor early.

This avoids some retry attemps from the other nodes at first startup.

* When we WARN about hostname, use IP address in HBA.

Rather than adding an hostname that we know faulty in the HBA file, we add
one of the IP addresses of the hostname instead. We might want to revisit
this (add all IP addresses maybe? or find the one we want to add by
connecting, like we do for the monitor?), but it allows the docker-compose
demo to just work with a minimum of trouble.

* Use PQExec() when paramCount is zero, allowing DDLs.

* User Postgres timing portability layer to track times.

* Implement a pg_autoctl do demo application.

The goal is to show what happens client-side (in the application) at
failover time. Handling disconnections properly and reconnecting to the same
database connection string to reach the new primary.

The demo application runs concurrent clients that each connect to a Postgres
database multi-host connection string, and INSERT some statistics about the
time taken to connect, and how many connection attempts where made.

* Add a demo container in the docker-compose setup.

* Accept that ping can be ok but then connection fail.

* Introduce a new connection type for applications.