Review pg_autoctl show state output.
Make it clear which kind of connection is given by which node, in particular
that when in the SINGLE or WAIT_PRIMARY state we are still delivering a
read-write connection to the application.

The "Reachable" column is now a "Connection" column and the node health as
seen by the monitor is encoded in the single-char *, !, or ?, depending on
the health value (1, 0, -1).