Do not leak psycopg2 connections during testing which can lead to flakyness
According to psycopg2 documentation connections are not closed on context exit.
This is in contrast with connection cursors that actually do so.

It is not an inherent problem in itself, however, calls to get_state() in a
loop, such as from wait_until_*, can overflow the connection limits and cause
the test suite to fail. This is particularly noticable when the system is slow
to respond, for example if it is run under valgrind.

This might address other objerved flakyness in the tests due to connection
exhaustion with messages as:
"psycopg2.OperationalError: FATAL:  sorry, too many clients already"