Remove deprecated IPDB code from the tests.
The IPDB module is deprecated already quite some time since
it has several architectural issues that make it almost
impossible to fix in some cases.

This PR replaces IPDB with its successor NDB.

At the same time all raw IPRoute calls were translated into
NDB calls, since the latter works in sync with the OS and
does not return from commit() before the changes are
actually applied to the system, unlike IPRoute, that returns
immediately not waiting for interfaces or addresses to be
set up.

This is just a translation, no code logic was affected
or refactored.