Plug connection leaks found during profiling (#582)
* Attempt to address review comments

* Plug connection leaks found during profiling

It seems that pgsql_execute_with_params() during its lifetime has been
inconsistently altered. The latest version notes in the comments that the
connection is not persistant to facilitate error handling. However that was not
entirely true and several parts of the code assumed it to not be true. Others
assumed to be true and failed to release the connection once used.

For the sake of clarity, the function will now explicitly close the connection
that has used, regardless of wether it is a new or existing connection. That
simplifies most of the code and plugs the connection leaks.

It also unconvers an inconsistency on the connections used for notification. The
code mixed the connection it was using to listen to events from the monitor and
with others. A new PGconn member has been added in the monitor struct to
distinguish between the two distinct cases.

* Rename emptyChannel to emptyChannelList as per review comments

* Perform an indent run

* Use monitor.notificationClient instead of pgsql where appropriate and close that connection when appropriate

* Address review comments.

 * Remove confusing pgsql_listen calls in favour of a new user friendly call
 * Use the same tight conn loop around monitor_wait_for_state_change everywhere
 * Correctly close the connections after it.

* Extention version functions require a multi statement connection

Addresses CI failure of test_extension_update case.

* Indent run

Co-authored-by: Georgios Kokolatos <gkokolatos@pm.com>