Allow registering a new node when there is no current primary.
It might be that we don't have a candidate for failover, and we lost the
primary. In that case we can register a new node from one of the standby
nodes and then see about promoting it.

At the moment we don't arrange for the new node to be automatically promoted
as a primary.