Fix a pipe issue and return code (#619)
* Fix a pipe issue and return code

Between closing the pipe files and waiting the program to exit,
the program may write to the closed pipe files, which causes
the program got the signal SIGPIPE and exit abnormally.
A safer way is to close these pipe files after the program does
exit.

Another issue is to extract the return code. `WEXITSTATUS(status)`
should be only employed if WIFEXITED returned true. Similarly,
if `WIFSIGNALED(status)` is true, set the signo as the return code.

* Update exit status: 128 + signo

Co-authored-by: Dimitri Fontaine <dimitri@citusdata.com>