Use PGDATA owner for the systemd service file. (#666)
The current user might be root or another interactive system user so that
it's possible to use sudo to write the systemd unit files. Typically, the
"postgres" user is not granted sudo privileges, and that's a good thing.

With this changes, if pg_autoctl show systemd is used from the root account
with a --pgdata created by the "postgres" user, we expect to obtain the
proper systemd unit file.