Review the Group FSM wait_primary decisions.
The code based some of the decision making on the candidate priority, which
is the wrong replication setting to consider here. To know if a node should
be wait_primary or primary, what's important is replication quorum enabled
standby nodes, and the fact that those are in the SECONDARY state.

There is then a special case when all the nodes are async, where we allow
the primary to be in the PRIMARY state as long as at least one node is in
the SECONDARY state, where in some cases it would otherwise be set to
WAIT_PRIMARY to disable sync rep. When all the nodes are async anyway, then
synchronous_standby_names is always computed to be ''.