Allow manual promotion of nodes with candidate priority zero. (#661)
* Allow manual promotion of nodes with candidate priority zero.

When promoting a node that is not setup as a candidate for failover, we
increment its candidate priority, perform the failover, and then reset its
candidate priority, as usual.

Allowing this mode of operation allows operating a pg_auto_failover cluster
in a fully manual way. When all the nodes have candidate priority set to
zero, it is still possible to manual trigger a promotion and target any
node.

This is a pre-requisite to being able to drop a node in all cases, including
when it's the primary node and there is no candidate for failover. Our
design is complete when it's always possible to run `pg_autoctl drop node`.

* Allow registering a new node when there is no current primary.

It might be that we don't have a candidate for failover, and we lost the
primary. In that case we can register a new node from one of the standby
nodes and then see about promoting it.

At the moment we don't arrange for the new node to be automatically promoted
as a primary.

* Fix pg_autoctl set node candidate-priority when skipping apply_settings.

* Allow setting replication quorum in more cases/states.

Handle setting replication quorum and APPLY_SETTINGS the same way that we're
doing with candidate priority.

* Fix comments in computing synchronous standby names.

* Add failover test coverage with zero candidates.

* Fix connection lost problem on the monitor.

* Review the Group FSM wait_primary decisions.

The code based some of the decision making on the candidate priority, which
is the wrong replication setting to consider here. To know if a node should
be wait_primary or primary, what's important is replication quorum enabled
standby nodes, and the fact that those are in the SECONDARY state.

There is then a special case when all the nodes are async, where we allow
the primary to be in the PRIMARY state as long as at least one node is in
the SECONDARY state, where in some cases it would otherwise be set to
WAIT_PRIMARY to disable sync rep. When all the nodes are async anyway, then
synchronous_standby_names is always computed to be ''.